Moving nexus Teams specific structs to Teams messaging_driver
